@@ -1,17 +1,20 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>IFToolsBriefings.Server</RootNamespace>
     <AssemblyName>IF-ToolsBriefings.Server</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.7.300" />
+    <PackageReference Include="AWSSDK.S3" Version="3.7.307.24" />
+    <PackageReference Include="dotenv.net" Version="3.1.3" />
     <PackageReference Include="EntityFramework" Version="6.4.4" />
-    <PackageReference Include="METAR.NET.Portable" Version="3.0.0-beta-1" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="5.0.4" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="6.0.0-preview.3.21201.2" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Metar.Decoder" Version="1.0.5" />
+    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="8.0.4" />
+    <PackageReference Include="MongoDB.EntityFrameworkCore" Version="7.0.0-preview.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
   <ItemGroup>
@@ -20,9 +23,13 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Content Update="appVersion.json">
+    <Content Update="version.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
+  </ItemGroup>
+
+  <ItemGroup>
+    <_ContentIncludedByDefault Remove="wwwroot\images\attachments\9ceb5432-d71c-40ce-90dc-6a4fe9f3c750.jpg" />
   </ItemGroup>
 
 
