@@ -15,18 +15,14 @@
             </NavLink>
         </li>
         <li class="nav-item px-3">
-            <ul class="nav nav-group">
-                <li class="nav-item px-3">
-                    <NavLink class="nav-link" href="/new" Match="NavLinkMatch.All">
-                        <span class="feather icon-file-plus" aria-hidden="true"></span> New Briefing
-                    </NavLink>
-                </li>
-                <li class="nav-item px-3">
-                    <NavLink class="nav-link" href="/find" Match="NavLinkMatch.All">
-                        <span class="feather icon-search" aria-hidden="true"></span> Find Briefing
-                    </NavLink>
-                </li>
-            </ul>
+            <NavLink class="nav-link" href="/new" Match="NavLinkMatch.All">
+                <span class="feather icon-file-plus" aria-hidden="true"></span> New Briefing
+            </NavLink>
+        </li>
+        <li class="nav-item px-3">
+            <NavLink class="nav-link" href="/find" Match="NavLinkMatch.All">
+                <span class="feather icon-search" aria-hidden="true"></span> Find Briefing
+            </NavLink>
         </li>
     </ul>
     <div class="nav-bottom-bar">
