@@ -4,11 +4,11 @@ namespace IFToolsBriefings.Shared.Data.Types
 {
     public class EditBriefingPostParameters
     {
-        public int Id { get; set; }
+        public string Id { get; set; }
         public Briefing EditedBriefing { get; set; }
         public string EditPassword { get; set; }
 
-        public EditBriefingPostParameters(int id, Briefing editedBriefing, string editPassword)
+        public EditBriefingPostParameters(string id, Briefing editedBriefing, string editPassword)
         {
             Id = id;
             EditedBriefing = editedBriefing;
